Improve logging and monitoring of the cwl-to-es
With this change, the cwl-to-es lambda function will log a single line per execution, with the result of that execution.

It will log in JSON format so it's easier to parse via a log metric filter.

I've also created a log metric filter to monitor the amount of errors thrown by the lambda function.

As per https://github.com/skyscrapers/engineering/issues/211